% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeIMG.R
\name{analyzeIMG}
\alias{analyzeIMG}
\title{Supply an image and use OpenAI to interpret the image.}
\usage{
analyzeIMG(
  analysis_image,
  AI_role,
  analysis_prompt,
  LLM,
  apiKey = apiKey,
  connecttimeout = 90
)
}
\arguments{
\item{analysis_image}{Either a URL for an image or a link to a locally stored image.}

\item{AI_role}{A character string defining how the LLM is to behave.}

\item{analysis_prompt}{A character string saying exactly what you are looking for in the image.}

\item{LLM}{A character string with the name of the model such as gpt-4o or gpt-3.5-turbo.}

\item{apiKey}{The OpenAI API key.}

\item{connecttimeout}{The number of seconds maximum wait for a response.}
}
\description{
Supply an image and use OpenAI to interpret the image.
}
\keyword{API,}
\keyword{GPT,}
\keyword{OpenAI,}
\keyword{images}
