% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeTXT.R
\name{analyzeTXT}
\alias{analyzeTXT}
\title{Supply text and use OpenAI to interpret the body of text.}
\usage{
analyzeTXT(
  analysis_text,
  AI_role,
  analysis_prompt,
  LLM,
  apiKey = apiKey,
  connecttimeout = 90
)
}
\arguments{
\item{analysis_text}{A character string with the text to be interpreted.}

\item{AI_role}{A character string defining how the LLM is to behave.}

\item{analysis_prompt}{A character string saying exactly what you are looking for in the text.}

\item{LLM}{A character string with the name of the model such as gpt-4o or gpt-3.5-turbo.}

\item{apiKey}{The OpenAI API key.}

\item{connecttimeout}{The number of seconds maximum wait for a response.}
}
\description{
Supply text and use OpenAI to interpret the body of text.
}
\keyword{API,}
\keyword{GPT,}
\keyword{OpenAI,}
\keyword{interpretation}
\keyword{text,}
