% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPDF.R
\name{extractPDF}
\alias{extractPDF}
\title{Supply a PDF that has text and retrieve the text for all or specific pages.}
\usage{
extractPDF(pdf_path, start_page = NULL, end_page = NULL)
}
\arguments{
\item{pdf_path}{The file and its location on the local computer.}

\item{start_page}{Value of the page number for the end of the extraction if a start_page specified.}
}
\description{
Supply a PDF that has text and retrieve the text for all or specific pages.
}
\keyword{PDF,}
\keyword{extract}
\keyword{range}
\keyword{text}
\keyword{text,}
