% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOAI.R
\name{queryOAI}
\alias{queryOAI}
\title{Supply a question and use OpenAI to answer the question by using its knowledge set.}
\usage{
queryOAI(prompt, LLM, temperature, apiKey = apiKey, connecttimeout = 90)
}
\arguments{
\item{prompt}{A character string with the question.}

\item{LLM}{A character string with the name of the model such as gpt-4o or gpt-3.5-turbo.}

\item{temperature}{A value from 0 to 1 that controls the fidelity of the response (1=best).}

\item{apiKey}{The OpenAI API key.}

\item{connecttimeout}{The number of seconds maximum wait for a response.}
}
\description{
Supply a question and use OpenAI to answer the question by using its knowledge set.
}
\keyword{API,}
\keyword{GPT,}
\keyword{OpenAI,}
\keyword{images}
